requirements.txt
numpy==2.2.4
flask-wtf==1.2.2
google-api-python-client==2.167.0
google-cloud-bigquery==3.31.0
pandas==2.2.3
plotly==6.0.1
pyarrow==19.0.1